Preserve select in table and enable/disable menu and tool bar item.
If beginResetModel is called then the selection on the table is
lost. To update the data and keep the selection it is necessary
to calls the begin(Insert|Remove)Row functions and the emit dataChanged.

Note: dataChnaged is not being responsed to by the QTableView at this
commit.

Added more debug methods for table-model and main-window/

Refactor the menu and tool bar setup code to hide
boiler plate code in helper functions. keep a list
of all the actions and the function that can set their
enable state.

Set the enable state each time the state may have changed.