Implement commit log viewer for a file
Use the same pattern as the diff viewer. (Fix dif viewer API to remove the parent that is not used.)

wb_git_project holds the code to walker the commit history and only keep the entries that are
related to the file.

Calculate the changed files for each commit.

Note: This code is not smart enough to notice a rename of a file yet.