Add preferences for UI font and Code font and use in the annotate and diff code stop using "font" as a variable name as it hides the font() Qt function.
Change svn log history to use the thread switch for slow svn calls,
but that required changing the way thread switcher code worked.

Rename all functions that use yield to add "_Bg" to their names to make
it obvious that thread switcher must be used with them. Use "yield from"
to call one _Bg function from another.

Add a thead switcher version of tableActionViewRepo and tableActionViewRepo.
tableActionViewRepo now takes a finalise function as result of the tableActionViewRepo_Bg
cannot be used in an if.