Implement rebase --interactive features
Suport reword, drop and squash from the log view.
Limits rebase to unpushed commits to avoid the issue of rewritting published history.

To use rebase --interactive need to over ride the GIT_SEQUENCE_EDITOR to allow
the rebase commands to be given to git and GIT_EDITOR to allow the new commit message
to be provided to git.

The existing askpass client and server have been extended to provide the sequence-editor
and editor callbacks.

Renamed the askpass code to be the callback code.
wb_git_project now holds the code to start the singleton server.

Prevent polluting the shell environment by using the GitPython udpate_environment()
mechanish for the repo object.

refactor: add commitMessageHeader() function

update build scripts to handle the help program name change from askpass
to callback.