Refactory the dialog feedback mechanism
WbFeedbackGridLayout can validate all the widgets inside the layout.
WbLineEdit takes a validator object that is one of the WbValidateLIneEdit
classes.
WbCHeckBoxLineEdit combines a checkbox with a lineedit widget and
honors the WbLineEdit validation interface.

wb_scm_project_dialogs and the svm, hg and git specific code is refactored
to use the new feedback API.